# Optimizations for writing binary strings.