## Description of changes

This PR optimizes the process by which the binary IonWriter writes strings.

### Performance change summary \[[benchmark source](https://gist.github.com/zslayton/e77d1298627645908c4c3d5445896879)\]

Performance is substantially improved for Strings small enough to benefit from buffer re-use. Strings that are too large to fit in the buffer see only a modest benefit.

#### Writing small strings (0 to 64 characters)

&nbsp; | Time Required | GC Time
| ----------- | :--------------: | :---------: |
| Before | 1,396.812 ± 95.650 ms/op | 5,771ms | 
| After | 804.164 ± 32.502 ms/op | 844ms | 
| &nbsp; | 42.43% faster | 85.38% lower |

#### Writing medium-sized strings (65 to 4096 characters)

&nbsp; | Time Required | GC Time
| ----------- | :--------------: | :---------: |
| Before | 569.213 ± 3.963 ms/op | 0ms | 
| After | 401.250 ± 23.278   ms/op | 0ms | 
| &nbsp; | 29.51% faster | 0% lower |

#### Writing large strings (4097 or more characters)

&nbsp; | Time Required | GC Time
| ----------- | :--------------: | :---------: |
| Before | 1,083.512 ± 41.708 ms/op | 455ms | 
| After | 1,031.493 ± 54.141   ms/op | 718ms | 
| &nbsp; | 4.80% faster | 42.2% higher |

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.